% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-docker-container.R
\name{container_stream_logs}
\alias{container_stream_logs}
\title{Stream logs for a container.}
\usage{
container_stream_logs(container)
}
\arguments{
\item{container}{A container object.}
}
\value{
A connection
}
\description{
Stream logs for a container.
}
