% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-docker.R
\name{docker_cmd_con}
\alias{docker_cmd_con}
\title{Run a docker command on a host, returning a connection.}
\usage{
docker_cmd_con(host, cmd = NULL, args = NULL, docker_opts = NULL, ...)
}
\arguments{
\item{host}{A host object.}

\item{cmd}{A docker command, such as "run" or "ps"}

\item{args}{Arguments to pass to the docker command}

\item{docker_opts}{Options to docker. These are things that come before the
docker command, when run on the command line.}

\item{...}{Other arguments passed to the SSH command for the host}
}
\description{
Run a docker command on a host, returning a connection.
}
