% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-mysql.R
\name{make_init_db_directory}
\alias{make_init_db_directory}
\title{Create MySQL initialisation directory.}
\usage{
make_init_db_directory(path_to_mysql_dump, username = NULL, password = NULL)
}
\arguments{
\item{path_to_mysql_dump}{Path to the MySQL dump file.}

\item{username}{Username}

\item{password}{Password, or NULL, if no password.}
}
\description{
Create a directory from which to initialise the database. This will be bound to
/docker-entrypoint-initdb.d.
}
\details{
This directory contains the MySQL dump file which is copied here, and if username is not null,
also a SQL file which creates the given user.
and grants them readonly privileges.
}
