% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-system.R
\name{mkdir_temporary}
\alias{mkdir_temporary}
\title{Create a temporary directory}
\usage{
mkdir_temporary(prefix = "resurrectr")
}
\arguments{
\item{prefix}{A character vector giving the initial part of the temporary directory name.}
}
\value{
A character vector of the canonical file path of the created temporary directory.
}
\description{
Creates a temporary directory, returning the file path of the freshly created directory.
This function stops on failure rather than merely showing warnings.
}
