% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-mysql.R
\name{resurrect_mysql}
\alias{resurrect_mysql}
\title{Resurrect a MySQL database from a dump file}
\usage{
resurrect_mysql(path_to_mysql_dump, version = "5.7", quiet = FALSE)
}
\arguments{
\item{path_to_mysql_dump}{Path to the MySQL dump file.}

\item{version}{Version of MySQL to run. In Docker language, this is a tag of the mysql image.
See https://hub.docker.com/r/library/mysql/tags/ for available versions.}

\item{quiet}{Whether container output is echoed.}
}
\value{
a resurrected_mysql object
}
\description{
Ideally, resurrectr should support different "drivers", eg Docker, Vagrant etc. However for now, only
Docker is supported.
}
