% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resurrected.R
\name{resurrected}
\alias{resurrected}
\alias{is.resurrected}
\title{Create a "resurrected" object}
\usage{
resurrected(subclass, finalizer = NULL, ...)

is.resurrected(x)
}
\arguments{
\item{subclass}{name of subclass. "resurrected" is an abstract base class, so you
must supply this value. `resurrected_` is automatically prepended to the
class name.}

\item{finalizer}{function to call on finalization or NULL. Function must accept
a single argument, which will be the object to finanlise.}

\item{...}{fields used by object.}

\item{x}{object to test for "resurrected"-ness.}
}
\description{
`resurrected()` is the standard constructor for resurrected-objects and `is.resurrected()` tests.
}
